pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "molecule-parser"
version = "0.1.0"
dependencies = [
  "biopython",
  "gradio",
]

[project.optional-dependencies]
dev = [
  "black"
]

[tool.black]
line-length = 88
target-version = ['py39']
